adapt edas call chains to kubernetes
