+++ modules/scheduler/executor/util/util.go
@@ -14,6 +14,8 @@
 package util
 
 import (
+	"encoding/base64"
+	"fmt"
 	"os"
 	"sort"
 	"strconv"
@@ -241,4 +243,64 @@ func IsNotFound(err error) bool {
 		return true
 	}
 	return false
+}
+
+// GetClient get http client with cluster info.
+func GetClient(clusterName string, manageConfig *apistructs.ManageConfig) (string, *httpclient.HTTPClient, error) {
+	inetPortal := "inet://"
+
+	hcOptions := []httpclient.OpOption{
+		httpclient.WithHTTPS(),
+	}
+
+	// check mange config type
+	switch manageConfig.Type {
+	case apistructs.ManageProxy, apistructs.ManageToken:
+		// cluster-agent -> (register) cluster-dialer -> (patch) cluster-manager
+		// -> (update) eventBox -> (update) scheduler -> scheduler reload executor
+		if manageConfig.Token == "" || manageConfig.Address == "" {
+			return "", nil, fmt.Errorf("token or address is empty")
+		}
+
+		hc := httpclient.New(hcOptions...)
+		hc.BearerTokenAuth(manageConfig.Token)
+
+		if manageConfig.Type == apistructs.ManageToken {
+			return manageConfig.Address, hc, nil
+		}
+
+		// parseInetAddr parse inet addr, will add proxy header in custom http request
+		return fmt.Sprintf("%s%s/%s", inetPortal, clusterName, manageConfig.Address), hc, nil
+	case apistructs.ManageCert:
+		if len(manageConfig.KeyData) == 0 ||
+			len(manageConfig.CertData) == 0 {
+			return "", nil, fmt.Errorf("cert or key is empty")
+		}
+
+		certBase64, err := base64.StdEncoding.DecodeString(manageConfig.CertData)
+		if err != nil {
+			return "", nil, err
+		}
+		keyBase64, err := base64.StdEncoding.DecodeString(manageConfig.KeyData)
+		if err != nil {
+			return "", nil, err
+		}
+
+		var certOption httpclient.OpOption
+
+		certOption = httpclient.WithHttpsCertFromJSON(certBase64, keyBase64, nil)
+
+		if len(manageConfig.CaData) != 0 {
+			caBase64, err := base64.StdEncoding.DecodeString(manageConfig.CaData)
+			if err != nil {
+				return "", nil, err
+			}
+			certOption = httpclient.WithHttpsCertFromJSON(certBase64, keyBase64, caBase64)
+		}
+		hcOptions = append(hcOptions, certOption)
+
+		return manageConfig.Address, httpclient.New(hcOptions...), nil
+	default:
+		return "", nil, fmt.Errorf("manage type is not support")
+	}
 }
