+++ modules/scheduler/impl/cluster/hook.go
@@ -41,6 +41,10 @@ var (
 	ErrClusterAddrNotSet = errors.New("cluster addr not set")
 )
 
+const (
+	REGISTRY_ADDR = "addon-registry.default.svc.cluster.local:5000"
+)
+
 // Hook receive clusterevent， reference `apistructs.ClusterEvent`
 func (c *ClusterImpl) Hook(clusterEvent *apistructs.ClusterEvent) error {
 	if !strutil.Equal(clusterEvent.Content.Type, clusterTypeDcos, true) {
@@ -395,8 +399,7 @@ func (c *ClusterImpl) createEdasExecutor(clusterEvent *apistructs.ClusterEvent) 
 			"CLUSTERID":       clusterEvent.Content.SchedConfig.ClusterID,
 			"REGIONID":        clusterEvent.Content.SchedConfig.RegionID,
 			"LOGICALREGIONID": clusterEvent.Content.SchedConfig.LogicalRegionID,
-			"KUBEADDR":        clusterEvent.Content.SchedConfig.K8sAddr,
-			"REGADDR":         clusterEvent.Content.SchedConfig.RegAddr,
+			"REGADDR":         REGISTRY_ADDR,
 		},
 	}
 	svcPath := strutil.Concat(clusterPrefix, clusterEvent.Content.Name, clusterEdasSuffix)
@@ -493,8 +496,9 @@ func patchEdasConfig(local *ClusterInfo, request *apistructs.ClusterInfo) error 
 		local.Options["CLUSTERID"] = request.SchedConfig.ClusterID
 		local.Options["REGIONID"] = request.SchedConfig.RegionID
 		local.Options["LOGICALREGIONID"] = request.SchedConfig.LogicalRegionID
-		local.Options["KUBEADDR"] = request.SchedConfig.K8sAddr
-		local.Options["REGADDR"] = request.SchedConfig.RegAddr
+		if local.Options["REGADDR"] == "" {
+			local.Options["REGADDR"] = REGISTRY_ADDR
+		}
 	}
 	return nil
 }
